+++ lib/Popups/photo_more_actions.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:memories_photos/Editor/editor_page.dart';
 import 'package:memories_photos/Structs/photo.dart';
+import 'package:memories_photos/Widgets/expressive_button.dart';
 import 'package:open_filex/open_filex.dart';
 
 Future<bool> photosMoreActionPopup(BuildContext context, Photo photo, bool moreActions) async {
@@ -12,24 +13,33 @@ Future<bool> photosMoreActionPopup(BuildContext context, Photo photo, bool moreA
       backgroundColor: Theme.of(context).colorScheme.secondaryContainer,
       child: Container(
         width: 340,
-        height: 360,
-        padding: .all(15),
+        height: moreActions ? 610 : 280,
+        padding: .symmetric(vertical: 15),
         child: ListView(
           physics: BouncingScrollPhysics(),
           children: [
-            moreActions ? ListTile(
-              title: Text("View in"),
-              leading: Icon(Icons.shortcut_outlined),
-              onTap: () {
+            ExpressiveButton.top(
+              text: "View in",
+              icon: Icon(Icons.shortcut_outlined),
+              onClick: () {
                 needsRefresh = true;
                 OpenFilex.open(photo.path);
                 Navigator.of(context).pop();
               },
+            ),
+            moreActions ? ExpressiveButton(
+              text: "Change Comment",
+              icon: Icon(Icons.insert_comment_outlined),
+              onClick: () async {
+                needsRefresh = true;
+                await photo.showEditCommentPopup(context, () {});
+                Navigator.of(context).pop();
+              },
             ) : SizedBox(),
-            moreActions ? ListTile(
-              title: Text("Details"),
-              leading: Icon(Icons.info_outline_rounded),
-              onTap: () {
+            moreActions ? ExpressiveButton(
+              text: "Details",
+              icon: Icon(Icons.park_outlined),
+              onClick: () {
                 needsRefresh = true;
                 Navigator.of(context).pop();
                 Navigator.of(context).push(MaterialPageRoute(builder: (context) => Dialog.fullscreen(
@@ -40,46 +50,37 @@ Future<bool> photosMoreActionPopup(BuildContext context, Photo photo, bool moreA
                 )));
               },
             ) : SizedBox(),
-            moreActions ? ListTile(
-              title: Text("Change Comment"),
-              leading: Icon(Icons.insert_comment_outlined),
-              onTap: () async {
-                needsRefresh = true;
-                await photo.showEditCommentPopup(context, () {});
-                Navigator.of(context).pop();
-              },
-            ) : SizedBox(),
-            ListTile(
-              title: Text("Edit"),
-              leading: Icon(Icons.edit_rounded),
-              onTap: () {
-                needsRefresh = true;
-                Navigator.of(context).pop();
-                Navigator.of(context).push(MaterialPageRoute(builder: (context) => EditorPage(photo: photo)));
-              },
-            ),
-            moreActions ? ListTile(
-              title: Text(photo.isInFavorites ? "Remove from favorites" : "Add to favorites"),
-              leading: Icon(photo.isInFavorites ? Icons.favorite_rounded : Icons.favorite_border_rounded),
-              onTap: () {
+            moreActions ? ExpressiveButton(
+              text: photo.isInFavorites ? "Remove from favorites" : "Add to favorites",
+              icon: Icon(photo.isInFavorites ? Icons.favorite_rounded : Icons.favorite_border_rounded),
+              onClick: () {
                 needsRefresh = true;
                 photo.addToFavorites(context);
                 Navigator.of(context).pop();
               },
             ) : SizedBox(),
-            moreActions ? ListTile(
-              title: Text("Delete"),
-              leading: Icon(Icons.delete_outline_rounded),
-              onTap: () async {
+            moreActions ? ExpressiveButton(
+              text: "Delete",
+              icon: Icon(Icons.delete_outline_rounded),
+              onClick: () async {
                 needsRefresh = true;
                 await photo.showDeletePopup(context, () {});
                 Navigator.of(context).pop();
               },
             ) : SizedBox(),
-            ListTile(
-              title: Text("Share"),
-              leading: Icon(Icons.share_rounded),
-              onTap: () {
+            ExpressiveButton(
+              text: "Edit",
+              icon: Icon(Icons.draw_rounded),
+              onClick: () {
+                needsRefresh = true;
+                Navigator.of(context).pop();
+                Navigator.of(context).push(MaterialPageRoute(builder: (context) => EditorPage(photo: photo)));
+              },
+            ),
+            ExpressiveButton.end(
+              text: "Share",
+              icon: Icon(Icons.share_rounded),
+              onClick: () {
                 photo.share();
                 Navigator.of(context).pop();
               },